+++ bootique-kotlin-example/src/main/java/io/bootique/kotlin/sample/KotlinConfig.kt
@@ -1,0 +1,13 @@
+package io.bootique.kotlin.sample
+
+import io.bootique.Bootique
+import io.bootique.kotlin.config.withKotlinConfig
+
+fun main(args: Array<String>) {
+    Bootique.app("--config=classpath:config.kts", "--config=classpath:config1.kts", "--server")
+        .withKotlinConfig()
+        .autoLoadModules()
+        .run()
+}
+
+data class TestConfig(val name: String, val products: List<String>)