+++ bootique-kotlin/src/main/java/io/bootique/kotlin/config/modules/FactoryDSL.kt
@@ -1,0 +1,11 @@
+package io.bootique.kotlin.config.modules
+
+/**
+ * Marker for configuration Builders
+ *
+ * @author Ibragimov Ruslan
+ * @since 0.1
+ */
+@DslMarker
+@Target(AnnotationTarget.TYPE)
+annotation class FactoryDSL