+++ RELEASE-NOTES.md
@@ -1,3 +1,5 @@
 ## 0.1
 
 * Initial release
+* Extension functions
+* Kotlin Configuration Script